Add docstring to dataset
