+++ mlbench_core/dataset/nlp/pytorch/wmt17/__init__.py
@@ -1,1 +1,2 @@
+from .collate import collate_batch
 from .wmt17_dictionary import Dictionary
